html/dom/render-blocking/render-blocked-apis-by-module-preload-link.tentative.html

<!doctype html>
<title>Certain APIs should not trigger while rendering is blocked by a preload link</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/test-render-blocking.js"></script>

<link id="module-preload" rel="modulepreload" blocking="render"
      href="support/dummy-1.mjs?pipe=trickle(d1)">

<div id="dummy">Lorem ipsum</div>

<script>
const preload = document.getElementById('module-preload');
test_render_blocked_apis(
    preload,
    async () => {
      // Attach the module script into document. It should load instantly due to preloading.
      const script = document.createElement('script');
      const scriptObserver = new LoadObserver(script);
      script.type = 'module';
      script.src = 'support/dummy-1.mjs?pipe=trickle(d1)';
      const beforeLoad = performance.now();
      document.body.appendChild(script);
      await scriptObserver.load;
      const afterLoad = performance.now();
      assert_equals(document.getElementById('dummy').textContent, '1',
                    'preloadedmodule script should be executed');
      assert_approx_equals(beforeLoad, afterLoad, epsilon,
                           'preloaded module script should load instantly');
    },
    'Render-blocking module script is preloaded and executed');
</script>
